merge_contents: try harder to preserve hardlinks.
If a set of intended hardlinks sprawl across multiple filesystems, previous
code would try only the first filesystem it was merged to.  The new
implementation tries a mildly brute force approach- it tracks each spot
it explicitly copied a hardlink candidate to, rather than hardlinking.

Now, it tries each of those previous candidates- ie, each filesystem that
was involved.  A smarter algorithm would check st_dev to find which candidate,
however that info isn't easily exposed in our current design, and it likely
would fail in the face of a overlay/unionfs.  The brute force approach here
is more than good enough in light of that.